+++ mangaki/mangaki/templates/base.html
@@ -70,13 +70,11 @@
                                 Mangas
                             </a>
                         </li>
-                    {% if user.is_authenticated %}
                         <li class="dropdown">
                             <a href="{% url 'reco' %}">
                                 Recommandations
                             </a>
                         </li>
-                    {% endif %}
                         <li class="dropdown">
                             <a href="{% url 'events' %}">
                                 Événements
